Boy howdy fox ears, now whatever ears! They now support transparency, and are mirrored correctly.
These are really something. 7 or so hours later, me and formuoli have figured out that needed to make each ear a separate Box, how to wrangle the opengl matrices,  learned how skins work and where transparency is, and that no one can see the bottom of your neck. 

Basically, the only downside to this whole thing is that it uses up a few pixels on the bottom of the neck hat, which is hardly visible to start with, but we have a patch in the works to fix this. 

This has been a really cool challenge for some bored programmers, I hope this can be used by everyone!