Overhauled ears to support transparency and proper mirroring (#250)
* Boy howdy fox ears, now whatever ears! They now support transparency and are mirrored correctly.

These are really something. 7 or so hours later, me and formuoli have figured out that needed to make each ear a separate Box, how to wrangle the OpenGL matrices,  learned how skins work and where transparency is, and that no one can see the bottom of your neck. 

Basically, the only downside to this whole thing is that it uses up a few pixels on the bottom of the neck hat, which is hardly visible to start with, but we have a patch in the works to fix this. 

This has been a really cool challenge for some bored programmers, I hope this can be used by everyone!

* Ear skins can now be loaded from any pair of locations on the skins file

This ended up being a lot of work to do, however now ears can be loaded from any 8x8 location on the skins file, pre or post 1.8. More explanation is in the LayerFoxEars file, and in the PR

* Reverting unnecessary gitignore changes

Co-authored-by: HeyZeer0 <21368488+HeyZeer0@users.noreply.github.com>